remove wxWidgets submodule -- all C/C++ submodules are kept in MuPDF/thirdparty/owemdjee/
TODO: come up with a slightly better organization of those -- the trouble is those 'research-level' (i.e. *take-a-look-at-probably-useeful*) and 'production-level' (i.e. *taken-a-look-at-and-deemed-useful*) do _overlap_, so there's no clear, hard, distinction between those categories. (and having the migratory path mapped onto more than 2 categories sounds an awaful lot like *administration-for-the-purpose-of-administration*.